Container 테스트 코드 수정 & 리팩토링
