+++ src/containers/ListContainer.test.jsx
@@ -1,21 +1,24 @@
 import { render } from '@testing-library/react';
+import { useSelector, useDispatch } from 'react-redux';
 
 import ListContainer from './ListContainer';
 
-describe('ListContainer', () => {
-  const handleClickDeleteTask = jest.fn();
+jest.mock('react-redux');
 
-  it('with tasks, titles should be rendered', () => {
-    const tasks = [
+describe('ListContainer', () => {
+  const dispatch = jest.fn();
+
+  useSelector.mockImplementation((selector) => selector({
+    tasks: [
       { id: 1, title: 'Task-1' },
       { id: 2, title: 'Task-2' },
-    ];
+    ],
+  }));
+  useDispatch.mockImplementation(() => dispatch);
 
+  it('with tasks, titles should be rendered', () => {
     const { getByText } = render((
-      <ListContainer
-        tasks={tasks}
-        onClickDeleteTask={handleClickDeleteTask}
-      />
+      <ListContainer />
     ));
 
     expect(getByText(/Task-1/)).not.toBeNull();
